+++ Sources/AssetGen/Infrastructure/Data/GoogleSheets/Fetcher/GoogleSheetDataLoader.swift
@@ -24,14 +24,3 @@ struct GoogleSheetDataLoader: SheetDataLoader {
     return sections
   }
 }
-
-extension GoogleSheetDataLoader {
-  static func make(with config: AssetGenConfig.Localization) -> GoogleSheetDataLoader {
-    let config = GoogleSheetsAPIConfig(apiKey: config.apiKey, sheetId: config.sheetId)
-    let requestBuilder = URLRequestBuilder(baseURLString: config.baseUrl)
-    let requestExecutor = APIRequestExecutor(requestBuilder: requestBuilder)
-    let fetcher = GoogleSheetsFetcher(config: config, requestExecutor: requestExecutor)
-    let dataLoader = GoogleSheetDataLoader(fetcher: fetcher)
-    return dataLoader
-  }
-}
