+++ Sources/AssetGen/Infrastructure/Data/GoogleSheets/Fetcher/GoogleSheetDataLoaderFactory.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+struct GoogleSheetDataLoaderFactory {
+  static func make(with config: AssetGenConfig.Localization) -> SheetDataLoader {
+    let config = GoogleSheetsAPIConfig(apiKey: config.apiKey, sheetId: config.sheetId)
+    let requestBuilder = URLRequestBuilder(baseURLString: config.baseUrl)
+    let requestExecutor = APIRequestExecutor(requestBuilder: requestBuilder)
+    let fetcher = GoogleSheetsFetcher(config: config, requestExecutor: requestExecutor)
+    let dataLoader = GoogleSheetDataLoader(fetcher: fetcher)
+    return dataLoader
+  }
+}
